refactor, WIP
