+++ doctalk/1_vector_db/.env
@@ -1,0 +1,3 @@
+PINECONE_API_KEY=
+PINECONE_ENV=
+OCTOAI_TOKEN=
