+++ doctalk/3_lambda/api-gateway.py
@@ -4,10 +4,10 @@ import boto3
 apigateway = boto3.client("apigateway")
 lambda_client = boto3.client("lambda")
 
-LAMBDA_ARN = "MyLambdaARN"
+LAMBDA_ARN = "MyLambdaARM" # FIXME: Update me once you've created the lambda!
 API_NAME = "MyAPI"
 STAGE_NAME = "prod"
-RESOURCE_PATH = "askllama"
+RESOURCE_PATH = "doctalk"
 HTTP_METHOD = "POST"
 
 # Create REST API
