+++ doctalk/3_lambda/template.yaml
