+++ doctalk/4_frontend/src/components/Footer.tsx
