+++ doctalk/4_frontend/src/components/Prompt.tsx
