+++ doctalk/4_frontend/src/styles/components/hero.css
